feat: весы графа могут принимать значения с плавающей точкой

+++ Assets/DijkstraAlgorithm.cs
@@ -18,7 +18,7 @@ namespace DijkstraAlgorithm
         /// M(i, j) характеризует расстояние от вершины i до j. Вес ребра M(i, i) равен нулю.
         /// Расстояние считается всегда положительным.
         /// </summary>
-        private int[,] graph;
+        private double[,] graph;
         /// <summary>
         /// Количество вершин графа.
         /// </summary>
@@ -27,7 +27,7 @@ namespace DijkstraAlgorithm
         /// Массив минимальных расстояний от заданной вершины до любой другой вершины.
         /// Представляется в виде [L(0), L(1), ..., L(N)], где L(i) - расстояние от заданной вершины до вершины i.
         /// </summary>
-        private int[] minimum_distance;
+        private double[] minimum_distance;
         /// <summary>
         /// Лист из индексов вершин, по которым можно построить минимальный путь из вершины i до вершины j.
         /// </summary>
@@ -45,9 +45,9 @@ namespace DijkstraAlgorithm
         /// <returns>
         /// Индекс искомой вершины.
         /// </returns>
-        private int MinimumDistance(int[] distance, bool[] shortestPathTreeSet)
+        private int MinimumDistance(double[] distance, bool[] shortestPathTreeSet)
         {
-            int min = int.MaxValue;
+            double min = double.MaxValue;
             int minIndex = 0;
 
             for (int v = 0; v < verticesCount; ++v)
@@ -68,24 +68,24 @@ namespace DijkstraAlgorithm
         /// <param name="initGraph">
         /// Исходный граф с весами.
         /// </param>
-        public Dijkstra(int[,] initGraph)
+        public Dijkstra(double[,] initGraph)
         {
             verticesCount = initGraph.GetLength(0);
 
             for (int vertice = 0; vertice < verticesCount; vertice++)
-                if (initGraph[vertice, vertice] != 0)
+                if (initGraph[vertice, vertice] != 0.0)
                     throw new System.ArgumentException("Диагональные вершины графа должны быть равны нулю.");
 
             for (int row = 0; row < verticesCount; row++)
                 for (int column = 0; column < verticesCount; column++)
-                    if (row != column && initGraph[row, column] < 0)
+                    if (row != column && initGraph[row, column] < 0.0)
                         throw new System.ArgumentException("Весы графа должны быть больше нуля.");
 
             if (initGraph.GetLength(0) != initGraph.GetLength(1))
                 throw new System.ArgumentException("Матрица с весами графа должна быть квадратной.");
 
             graph = initGraph;
-            minimum_distance = new int[verticesCount];
+            minimum_distance = new double[verticesCount];
             VerticesPath = new List<int>();
         }
 
@@ -98,7 +98,7 @@ namespace DijkstraAlgorithm
         /// <returns>
         /// Минимальное расстояние L(vertice) от заданной вершины.
         /// </returns>
-        public int GetMinimumDistance(int vertice)
+        public double GetMinimumDistance(int vertice)
         {
             if (vertice < 0 || vertice >= verticesCount)
                 throw new System.ArgumentOutOfRangeException("Выбрана вершина вне диапазона имеющихся вершин.");
@@ -131,12 +131,12 @@ namespace DijkstraAlgorithm
             if (source < 0 || source >= verticesCount || endVertice < 0 || endVertice >= verticesCount)
                 throw new System.ArgumentOutOfRangeException("Выбрана вершина вне диапазона имеющихся вершин.");
 
-            int[] distance = new int[verticesCount];
+            double[] distance = new double[verticesCount];
             bool[] shortestPathTreeSet = new bool[verticesCount];
 
             for (int i = 0; i < verticesCount; ++i)
             {
-                distance[i] = int.MaxValue;
+                distance[i] = double.MaxValue;
                 shortestPathTreeSet[i] = false;
             }
 
@@ -158,7 +158,7 @@ namespace DijkstraAlgorithm
                 shortestPathTreeSet[u] = true;
 
                 for (int v = 0; v < verticesCount; ++v)
-                    if (!shortestPathTreeSet[v] && Convert.ToBoolean(graph[u, v]) && distance[u] != int.MaxValue && distance[u] + graph[u, v] < distance[v])
+                    if (!shortestPathTreeSet[v] && Convert.ToBoolean(graph[u, v]) && distance[u] != double.MaxValue && distance[u] + graph[u, v] < distance[v])
                         distance[v] = distance[u] + graph[u, v];
             }
